added recently extra

+++ src/test/java/steps/Hooks.java
@@ -2,6 +2,7 @@ package steps;
 
 import io.cucumber.java.After;
 import io.cucumber.java.Before;
+import io.cucumber.java.Scenario;
 import io.cucumber.java.bs.A;
 import org.openqa.selenium.WebDriver;
 import utilities.Driver;
@@ -9,18 +10,23 @@ import utilities.Driver;
 import java.util.concurrent.TimeUnit;
 
 public class Hooks {
-    WebDriver driver= Driver.getDriver();
+    WebDriver driver;
     @Before
-    public void setup(){
+    public void setup(Scenario scenario){
+       if (!scenario.getSourceTagNames().contains("@api")){
+           driver=Driver.getDriver();
         driver.manage().window().maximize();
         driver.manage().timeouts().implicitlyWait(10, TimeUnit.SECONDS);
         driver.manage().deleteAllCookies();
         System.out.println("Before Scenario Method");
-    }
+    }  }
     @After
-    public void tearDown() throws InterruptedException {
+    public void tearDown(Scenario scenario) throws InterruptedException {
+        if (!scenario.getSourceTagNames().contains("@api")) {
+
         Thread.sleep(3000);
         driver.quit();
         System.out.println("After Scenario Method");
     }
+}
 }
